Windows fixes
git-svn-id: svn+ssh://192.168.1.105/Users/mikael/icl-repository/trunk@597 763894fe-7313-0410-9b4e-9763914fbcec

+++ util/ic_hashtable.c
@@ -38,7 +38,7 @@ static const unsigned int primes[] = {
 805306457, 1610612741
 };
 const unsigned int prime_table_length = sizeof(primes)/sizeof(primes[0]);
-const float max_load_factor = 0.65;
+const double max_load_factor = 0.65;
 
 int
 ic_keys_equal_str(void *ptr1, void *ptr2)
